Removed: usage of picklistService and defined the endpoints in UtilService(#1wut90a)
