+++ src/services/UtilService.ts
@@ -1,4 +1,5 @@
-import { api } from '@/adapter';
+import { api, client } from '@/adapter';
+import store from '@/store';
 
 const fetchShipmentMethods = async (query: any): Promise <any>  => {
   return api({
@@ -8,6 +9,29 @@ const fetchShipmentMethods = async (query: any): Promise <any>  => {
   });
 }
 
+const getAvailablePickers = async (query: any): Promise <any> => {
+  return api({
+    url: 'performFind',
+    method: 'POST',
+    data: query,
+    cache: true
+  })
+}
+
+const createPicklist = async (query: any): Promise <any> => {
+  let baseURL = store.getters['user/getInstanceUrl'];
+  baseURL = baseURL && baseURL.startsWith('http') ? baseURL : `https://${baseURL}.hotwax.io/api/`;
+  return client({
+    url: 'createPicklist',
+    method: 'POST',
+    data: query,
+    baseURL,
+    headers: { "Content-Type": "multipart/form-data" },
+  })
+}
+
 export const UtilService = {
-  fetchShipmentMethods
+  createPicklist,
+  fetchShipmentMethods,
+  getAvailablePickers
 }
